move formatting to dedicated module
