+++ zotpilot/retrieval.py
@@ -1,4 +1,3 @@
-import re
 from typing import Any
 
 import torch
@@ -41,38 +40,3 @@ def similarity_search(
         )
 
     return results
-
-
-def format_context(results: list[dict[str, Any]], include_metadata: bool = True) -> str:
-    """
-    Format retrieved chunks into a context string for the LLM.
-
-    Args:
-        results: List of retrieval results
-        include_metadata: Whether to include metadata in the context
-
-    Returns:
-        Formatted context string
-    """
-    context_parts = []
-
-    for i, result in enumerate(results):
-        if include_metadata:
-            context_parts.append(
-                f"[CHUNK {i + 1} (Page {result['metadata']['page']})]:\n{result['text']}\n"
-            )
-        else:
-            context_parts.append(result["text"])
-
-    return "\n\n".join(context_parts)
-
-
-def format_response_with_citations(response: str) -> str:
-    """Format the response with highlighted citations."""
-    citation_pattern = r"\[(\d+(?:,\s*\d+)*)\]"
-    formatted_response = re.sub(
-        citation_pattern,
-        lambda m: f'<span style="background-color: #e6f3ff; padding: 1px 4px; border-radius: 3px;">{m.group(0)}</span>',
-        response,
-    )
-    return formatted_response
